console: Refactor webhook edit view use update action instead of direct api call
